chore: 리뷰 사항 반영
